+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import React, { useEffect } from 'react';
 import Header from './components/header';
-import Mission from './components/mission';
+import Missions from './components/missions';
 import Profile from './components/profile';
 import Rockets from './components/rockets';
 import './sass/App.css';
@@ -23,7 +23,7 @@ function App() {
         <Header />
         <Switch>
           <Route path="/" exact render={() => <Rockets store={store} />} />
-          <Route path="/mission" component={Mission} />
+          <Route path="/mission" component={Missions} />
           <Route path="/profile" component={Profile} />
         </Switch>
       </BrowserRouter>
