+++ src/components/common/mission.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Mission = () => (
+  <tr>
+    <th scope="row">Thaicom</th>
+    <td>Conubia. In. Lobortis luctus cum maecenas, consequat tincidunt Nulla nunc consequat. Sociosqu torquent eros Eget convallis maecenas. Risus aliquam varius hac lobortis. Sodales parturient et justo egestas tortor massa litora netus ad massa nibh. Vitae vel Curae; dolor tellus semper tincidunt fusce cubilia convallis habitant tristique consectetuer tempus, ut platea.</td>
+    <td><p className="ubadge">NOT A MENBER</p></td>
+    <td><button className="mission-button exit">Leave Mission</button></td>
+  </tr>
+);
+export default Mission;
